requirements.txt
torch
transformers
Pillow
tqdm
nltk
matplotlib
pandas
